Separate leaderboards page into components
