+++ next.config.js
@@ -7,7 +7,6 @@ const nextConfig = {
       {
         protocol: 'https',
         hostname: 'cloudflare-ipfs.com',
-        port: '',
         pathname: '/ipfs/**'
       }
     ]
